requirements.txt
flask
mysql-connector-python
pyjwt
werkzeug
